Run Neon NTT through SLOTHY and add Makefile
This adds a Makefile that runs the Neon NTT through SLOTHY. To accomodate this
the clean assembly is moved to dev/aarch64_clean/, while the mldsa/native/aarch64
contains the optimized assembly.

The main difference to mlkem-native is that we need set an explicit timeout
as optimizing the second loop doesn't result reasonable performance, but a good
solution is found within one minute on my Apple M4. I set the timeout to 2
minutes with the hope that it works on most platforms. We have have to
increase that later.

For now the clean backend is not tested in CI - that's left for a follow-up PR.
SLOTHY is also not run in CI, yet.
We probably want to put the assembly simplification scripts in place so we
can follow the same structure as in mlkem-native.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>