CI: Test liboqs integration
This commit adds CI to test that our liboqs integration does not break upon
changes in mldsa-native.
It currently works with the liboqs fork in
https://github.com/mkannwischer/liboqs which contains the changes required to
liboqs.
It performs an import with the current commit, builds liboqs, and runs the
tests for ML-DSA.

Ported and adapted from mlkem-native.
Unfortunately, our AVX2 backend results in AVX512 instructions which are not
supported by valgrind. This PR explicitly compiler for Haswell instead.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>