Add "internal" ACVP tests
This commit adds the tests for the internal signing and verification functions
from the ACVP testvectors.
There is a small inconvienence: We do not follow FIPS204 exactly in the
function signature for the internal functions as we take a prefix pre and a
message msg separately instead of receiving a formatted message including both
the prefix and the message.

I initially considered changing the APIs to align with FIPS204 but that is
undesirable as the message has variable-length and consequently the higher
level APIs would have to either allocate a variable length array for the
formated message or perform the computation of the hash mu themselves.
Hence, I decided to stick with the existing API.

To make the ACVP tests that have the full formatted message instead of
pre + msg separately, we can instead pass the formatted message as the message
together with an empty prefix pre.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>