CBMC: Re-model preconditions allowing NULL pointers
Previously if a pointer was allowed to be NULL, we would write
requires((p == NULL && len == 0) || memory_no_alias(p, len)).
This commit refactores that to
requires(len == 0 || memory_no_alias(p, len))
which feels more natural as there is no need to force p to NULL if len is 0.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>