ACVP Client: Adjust to handle deterministic test cases better
Before the ACVP client would implement tests for the determinstic ML-DSA
variant by passing an all-zero rnd to the the test harness. This felt a
bit hacky.
This commit changes it to handle the deterministic variant in the test harness
directly by adding sigGenDeterministic, sigGenInternalDeterministic,
sigGenPreHashDeterministic, and sigGenPreHashShake256Deterministic commands.

To make simplify it slightly in the acvp_client, the rnd argument is moved
to the end of each command.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>