native chknorm: 0/0xFFFFFFFF -> 0/1
Previously the native chknorm functions would follow the C implementation,
i.e., return 0 if all coefficients are within bound and 0xFFFFFFFF otherwise.
This leads to problems with the run-time dispatch in
https://github.com/pq-code-package/mldsa-native/pull/607
as we want to use -1 to signal that the current platforms lacks the required
capabilities to run the native code, and we should fall back to the C
implementation.

This commit changes the backend API to return 1 in the failure mode.
This will allow to implement the run-time dispatch.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>