Hoist domain separation logic into helper function
This commit introduces a helper function

```
  mld_prepare_domain_separation_prefix
```

for the preparation of the domain separation prefix.

This is used both for Pure ML-DSA and HashML-DSA and was previously
done in an ad-hoc manner for Pure ML-DSA.

The helper is part of the external API in order to facilitate
incremental APIs being built atop the APIs exposed by mldsa-native.

A CBMC proof is added.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>