SLOTHY: Superoptimize ntt.S for the Neoverse N1
This commit adds the optimized backend dev/aarch64_opt. For now this backend
only differs from the clean backend in the NTT which is superoptimized using
SLOTHY for the Neoverse N1. For all other files it's a simple copy of the clean
backend. A Makefile is added that performs the optimization.
CI is adjusted to test both the clean and the opt backend.

The first loop of the NTT can be optimized, in one go. The second loop is
too large (>350 instructions) and we, hence, use the split heuristic.

The autogen script is extended to allow running the optimization through the
--slothy flag.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>