ACVP Client: Do not rely on truncated SHA512 support in hashlib
Python 3.7's hashlib does not have support for truncated SHA512
(hashlib.new("sha512_224", ...) and hashlib.new("sha512_256", ...).
It seems it got added in 3.8, but the documentation does not mention it at all.

This commit removes the relying on both sha512_224 and sha512_256 support in
hashlib, and instead just uses sha512 and performs the truncation afterwards.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>