verify: Switch to constant-time memcmp
In the very end of verify one has to compare the input challenge to the
re-computed challenge. If they are equal (and some previous checks on
h and z passed), the signature is valid.
Currently, our constant-time tests do not declassify the message and we, hence,
need to declassify in this final step.
Before thi commit, the declassification would happen on the recomputed challenge
just before the memcmp.
Now that a constant-time memcmp was added in
https://github.com/pq-code-package/mldsa-native/pull/714, we might as well
use that; that plus a constant-time selections allows us to not use any
declassifications in verification, i.e., we do not leak the verification
result through timing.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>