Constant-time: Make signature declassifications explicit in verification
Previously, the signature would be declassified in the end of
signing (which is the most common threat model).
However, for rare use cases the signature may be secret in which
case it is useful to know which part of the code has to be
adapted.

This commit removes the declassification from the end of signing,
explicitly marks the signature as secret in the constant time tests
(to also classify the previously declassified challenge c), and
then adds necessary declassifications in verifiation.

We may consider eliminating some of those declassifications later.
In particular, the data dependencies in use_hint are not hard to
eliminate (in fact, the native implementations are already constant
time). The final comparison of the challenges can also easily
be turned into a constant-time comparison.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>